Implement the Preemptive Multitasking
Add schedule mechanism to switch from task to task. Replace the enter_user_mode with this scheduling.
Fix the masking "|=" in the APIC Timer to be "=" for the masking of interrupt. And set TPR to be 0, which avoids swallowing the timer's interrupt signals.
Fix the EOI in the `timer_handler` in `timer.c`, because the called `schedule()` also uses interrupts.