Implement sys_waitpid and zombie processes
Struct Task is updated with a new state - TASK_ZOMBIE, and a field parent to wake it up.
A task, when exit, is not destroyed immediately, but changes state to zombie, retain the return value (exit code), wakes up its parent if any, and let the parent collects the exit code, destroy the task.
A task, when sys_waitpid a child, it is blocked, until its child wakes it up, checks the child's state if it's zombie or not, then destroy it if yes.